+++ sync-for-commerce/src/codat_sync_for_commerce/models/operations/__init__.py
@@ -51,6 +51,7 @@ from .update_visible_accounts_sync_flow import (
     UpdateVisibleAccountsSyncFlowRequestTypedDict,
 )
 
+
 __all__ = [
     "CreateConnectionRequest",
     "CreateConnectionRequestBody",